refactor: 実装から entity への参照を Public API 経由に統一し、自己参照を lint で禁止 (FRESTYLE-165)
独立レビューで見つかった 2 点への対応。振る舞いは変えていない。

1) barrel を作ったのに実装側が Slice 内部を直接 import していた(25 ファイル)。
   Public API を通すという Phase の目的が未達で、index.ts のコメント
   「内部ファイルを直接 import してはいけない」と実態が食い違っていた。
   audit / learning-report / member / invitation の barrel は 1 つも
   barrel 経由で使われていない状態だった。実装ファイルを @/entities/<slice>
   経由に統一し、barrel が default を持たないため default import は
   名前付きへ変換した。

   テストは deep import のまま残す。vi.mock は対象モジュール自身を指す必要が
   あり、テストで barrel を読むと無関係な兄弟モジュールの破損で落ちるうえ、
   カバレッジの分母も膨らむため。

2) 境界 lint がテストを対象外にしていたため、Slice の自己参照が無警告だった。
   層間ルールがテストを除外しているのは「テストが上位層の Provider で包むのは
   正当」だからだが、自己参照には正当なケースが無い。Slice ごとの自己参照禁止
   ルールを追加し、テストにも適用した。追加した途端に未修正の 1 件
   (ProfileStatsSection.test.tsx)を検出したので相対パスに直した。

   プローブで (1)テストからの自 Slice 参照は検出 (2)他 Slice の barrel 参照は
   テストでは許可のまま、を実測で確認した。

あわせて globalIgnores に coverage を追加(生成物が lint 対象に入っていた)。
