+++ frontend/src/app/providers/AuthInitializer.tsx
@@ -2,7 +2,7 @@ import { ReactNode, useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { setAuthData, clearAuth, finishLoading } from '@/entities/user';
 import type { RootState } from '@/store';
-import authRepository from '@/entities/user/api/authRepository';
+import { AuthRepository as authRepository } from '@/entities/user';
 import Loading from '@/shared/ui/Loading';
 
 interface AuthInitializerProps {
