+++ frontend/src/hooks/useAuditLog.ts
@@ -1,5 +1,5 @@
 import { useEffect, useState } from 'react';
-import { AuditRepository, AuditEvent } from '@/entities/audit/api/auditRepository';
+import { AuditRepository, AuditEvent } from '@/entities/audit';
 
 /**
  * useAuditLog — 監査ログ（super_admin）を取得するフック。
