+++ frontend/src/hooks/useLoginCallback.ts
@@ -2,7 +2,7 @@ import { useEffect } from 'react';
 import { useSearchParams, useNavigate } from 'react-router-dom';
 import { useDispatch } from 'react-redux';
 import { setAuthData } from '@/entities/user';
-import authRepository from '@/entities/user/api/authRepository';
+import { AuthRepository as authRepository } from '@/entities/user';
 import { consumeInvitationToken } from '../lib/invitationToken';
 import { classifyApiError, getApiError } from '@/shared/lib/classifyApiError';
