+++ frontend/src/pages/admin-members/ui/AdminMembersPage.tsx
@@ -5,7 +5,7 @@ import type { RootState } from '@/store';
 import Loading from '@/shared/ui/Loading';
 import PageIntro from '@/shared/ui/PageIntro';
 import { useAdminMembers } from '@/hooks/useAdminMembers';
-import { Member } from '@/entities/member/api/adminMemberRepository';
+import { Member } from '@/entities/member';
 
 // AI 利用可否の 3 状態 ↔ select の値。
 function aiValue(m: Member): 'inherit' | 'on' | 'off' {
